refactor(cli): remove meta file support
Meta file was going to only be used for keeping list of when the
text-to-speech pre-processing script was run so that it knows what
files changed since the last run.

But I realized that git tags already exist, and should be used exactly
for this purpose, and will make my life simpler.