chore(Migration): temporary add migration scripts
I migrated my articles storage format from a long file to a Text Hoarder
repository. Previous format not including any metadata made things
difficult, but I did my best to get back the previous URL and time
accessed from the browser history. For pages that didn't match browser
history, I inferred the date accessed, and Googled the snippet to find
a likely URL.

Adding the migration scripts to this repository and would remove them
in the previous commit just so that they are here in the history.