fix(ReaderMode): prevent duplicate auto-trigger
Don't like the solution, but that's the best I could come up with.
The other options (sending a message OR separate bundles) seeded worse.

I also tried having a small bundle that is called in the autoTrigger
case that would define a global symbol and then import the regular
bundle, but I didn't know if async loading would be slower than the
overhead of waiting for storage.

If I really want to reduce latency, could consider fetching the page
HTML from the back end rather than wait for page to load.