chore(txt-to-md-migration): add migration script
Before text hoarder, all my saved articles were stored as .txt.

I used this migration script to convert them to .md by fetching the
content from the website again, and converting to to markdown if it is
close enough to the original text content. Otherwise, original text
content was converted to markdown.