fix(SaveText): fix duplicate save file request
Happens only after first sign in.
When "github" changes from "undefined" to not, the useAsyncState resets
the "existingFile" state from false to undefined, only to have it be
returned back to false.
The "SaveText" is rendered only while "existingFile" is not undefined.
Thus, this brief change in state causes the "SaveText" to re-initialized
twice, sending the duplicate saveFile request, and causing user to see
a false "this article has already been saved" message

Fixed by not using useAsyncState to avoid useAsyncState's logic of
resetting the state to undefined while new state is being loaded