feat(Stats): use Intl.Segmenter API
Was disappointed at how slot is - it takes 8 min to compute the stats
on my text-hoarder-store when using Intl.Segmenter, compared to 40 sec
using regex. What is more, word count between the two counts is almost
identical. Sentence count is undercoated in Regex by 20%, but even that
is not worth the giant slowdown.

Thus, the option to use Intl.Segmenter is behind a flag, rather than on
by default (not to mention that that API is not in Firefox yet)