fix(cli): correctly find year in the path
To allow Text Hoarder CLI to be invoked on a sub-directory of a Text
Hoarder repository, and yet still correctly infer the full URL for a
file based on the relative/absolute file path, I was using a simple
check to find the occurrence of `/2`` in the relative/absolute path.

However, this caused false positives in a URL that includes a year in
it, for example `2020/aaronparecki.com/2020/08/31/9/too-many-projects.md`

Since the relative path above does not begin with `/2` (it begins with
`2`), the path got incorrectly trimmed of at
`2020/08/31/9/too-many-projects.md`, which then caused the URL parsing
to crash.

Fixed by using Regex to find ^ or / followed by year